Revert "fixed somebugs"
