+++ src/styles/blocks/header/__container/header__containter.css
@@ -13,7 +13,7 @@
   }
 }
 
-@media screen and (max-width: 618px) {
+@media screen and (max-width: 606px) {
   .header__container {
     max-width: 282px;
     padding-top: 28px;
