+++ src/styles/blocks/header/header.css
@@ -1,20 +1,20 @@
 .header {
-  width: 100%;
-  box-sizing: border-box;
-  max-width: 882px;
-  min-height: 74px;
-  border-bottom: 1px solid rgba(84, 84, 84, .7);
-  margin: 0 auto;
-  margin-bottom: 40px;
-  display: flex;
-  justify-content: space-between;
-  flex-wrap: wrap;
-  padding-bottom: 41px;
+    width: 100%;
+    box-sizing: border-box;
+    max-width: 882px;
+    min-height: 74px;
+    border-bottom: 1px solid rgba(84, 84, 84, .7);
+    margin: 0 auto;
+    margin-bottom: 40px;
+    display: flex;
+    justify-content: space-between;
+    flex-wrap: wrap;
 }
 
 @media screen and (max-width: 635px) {
-  .header {
-    min-height: 56px;
-    padding-bottom: 30px;
-  }
+    .header {
+
+        min-height: 56px;
+        padding-bottom: 30px;
+    }
 }
